Diffusion with absorbtion. Return value fixups.
At the end of the computation a lot of useful data had been lost
(e.g. sample paths) that could be used for visualization during the
presentation. I made it available for reuse.

Finding maximal value of pdf in `solve_*` resulted in code duplication.
That part of the code has been moved to `sample_rejection.m` -- this
made `top` an optional argument.